Extract PDF import AI utilities
