+++ js/pdf-import.js
@@ -5,13 +5,19 @@ import { state } from './state.js';
 import { MARKER_SCHEMA, SPECIALTY_MARKER_DEFS, calculateCost, trackUsage } from './schema.js';
 import { showNotification, showConfirmDialog, isDebugMode, isPIIReviewEnabled, hashString } from './utils.js';
 import { saveImportedData } from './data.js';
-import { callClaudeAPI, hasAIProvider, getAIProvider, getActiveModelId, AI_IMPORT_REQUEST_TIMEOUT_MS } from './api.js';
+import { hasAIProvider, getAIProvider, getActiveModelId, AI_IMPORT_REQUEST_TIMEOUT_MS } from './api.js';
 import { obfuscatePDFText, sanitizeWithOllama, sanitizeWithOllamaStreaming, checkOllamaPII, reviewPIIBeforeSend } from './pii.js';
 import { getPdfDocument } from './pdfjs-loader.js';
 import { getProfileLocation, getActiveProfileId } from './profile.js';
 import { findOrCreateLabEntry } from './lab-entry-mutations.js';
 import { setLabEntryMarker, syncLabEntryInsulinMirror } from './lab-entry.js';
 import { closeModalOverlay, openModalOverlay } from './modal-lifecycle.js';
+import {
+  callImportAIWithStreamFallback,
+  formatImportError,
+  getUsageTokens,
+  tryParseJSON,
+} from './pdf-import-ai-utils.js';
 import { extractXLSXText, isCsvTextFile, isTextImportFile, isXlsxFile } from './pdf-import-spreadsheet.js';
 import { runPreflightChecks } from './pdf-import-preflight.js';
 import { normalizeParsedImportMarkers } from './pdf-import-marker-normalization.js';
@@ -64,12 +70,12 @@ import {
  *   handleMtDNAFile?: (file: File) => void | Promise<void>,
  *   handleDNAFile: (file: File) => void | Promise<void>,
  * }} PdfImportWindowHooks
- * @typedef {{ inputTokens?: number, outputTokens?: number }} ImportUsage
  */
 
 const pdfImportWindow = /** @type {Window & typeof globalThis & PdfImportWindowHooks} */ (window);
 
 export { buildMarkerReference, reconcileImportMarkerMappings } from './pdf-import-marker-mapping.js';
+export { tryParseJSON } from './pdf-import-ai-utils.js';
 export { extractXLSXText } from './pdf-import-spreadsheet.js';
 export {
   showImportPreview,
@@ -220,84 +226,6 @@ async function readFileArrayBuffer(file) {
       reader.onabort = () => reject(firstError);
       reader.readAsArrayBuffer(file);
     });
-  }
-}
-
-function isAIStreamAbortError(err) {
-  const message = String(err?.message || err || '').toLowerCase();
-  const name = String(err?.name || '').toLowerCase();
-  return message.includes('bodystreambuffer was aborted')
-    || message.includes('aborted by user')
-    || (message.includes('body') && message.includes('stream') && message.includes('abort'))
-    || name === 'aborterror';
-}
-
-async function callImportAIWithStreamFallback(request, label) {
-  try {
-    return await callClaudeAPI(request);
-  } catch (err) {
-    if (!request.onStream || request.signal?.aborted || !isAIStreamAbortError(err)) throw err;
-    if (isDebugMode()) console.warn(`[Import] ${label} stream aborted; retrying without streaming`, err);
-    try {
-      return await callClaudeAPI({ ...request, onStream: undefined, forceNonStream: true, requestTimeoutMs: AI_IMPORT_REQUEST_TIMEOUT_MS });
-    } catch (retryErr) {
-      if (isAIStreamAbortError(retryErr)) {
-        throw new Error('AI analysis request was aborted after retrying without streaming. The PDF text extracted correctly; try another model/provider if this persists.');
-      }
-      throw retryErr;
-    }
-  }
-}
-
-function formatImportError(err) {
-  if (isAIStreamAbortError(err)) {
-    return 'AI analysis request was interrupted after privacy review. Try again, or switch provider/model if it repeats.';
-  }
-  return err?.message || String(err);
-}
-
-function getUsageTokens(usage) {
-  const u = /** @type {ImportUsage} */ (usage || {});
-  return {
-    inputTokens: Number(u.inputTokens) || 0,
-    outputTokens: Number(u.outputTokens) || 0,
-  };
-}
-
-export function tryParseJSON(str) {
-  try { return JSON.parse(str); } catch {}
-  // Try trimming to last complete object (handles truncated output)
-  const lastBrace = str.lastIndexOf('}');
-  if (lastBrace > 0 && lastBrace < str.length - 1) {
-    try { return JSON.parse(str.slice(0, lastBrace + 1)); } catch {}
-  }
-  // Attempt to repair truncated JSON from local models
-  let s = str;
-  // Close any unterminated string
-  const quotes = (s.match(/"/g) || []).length;
-  if (quotes % 2 !== 0) s += '"';
-  // Try closing open arrays and objects
-  const opens = { '{': 0, '[': 0 };
-  let inString = false;
-  for (let i = 0; i < s.length; i++) {
-    if (s[i] === '"' && (i === 0 || s[i - 1] !== '\\')) { inString = !inString; continue; }
-    if (inString) continue;
-    if (s[i] === '{') opens['{']++;
-    if (s[i] === '}') opens['{']--;
-    if (s[i] === '[') opens['[']++;
-    if (s[i] === ']') opens['[']--;
-  }
-  // Remove trailing comma before closing
-  s = s.replace(/,\s*$/, '');
-  // Close unclosed brackets/braces
-  for (let i = 0; i < opens['[']; i++) s += ']';
-  for (let i = 0; i < opens['{']; i++) s += '}';
-  try {
-    const result = JSON.parse(s);
-    if (isDebugMode()) console.log('[PDF Parse] Repaired truncated JSON from model');
-    return result;
-  } catch (e2) {
-    throw new Error(`Model returned invalid JSON that could not be repaired. Try a more capable model.`);
   }
 }
