+++ service-worker.js
@@ -125,6 +125,7 @@ const APP_SHELL = [
   '/js/pdf-import-spreadsheet.js',
   '/js/pdf-import-preflight.js',
   '/js/pdf-import-progress.js',
+  '/js/pdf-import-ai-utils.js',
   '/js/pdf-import-review.js',
   '/js/pdf-import-marker-mapping.js',
   '/js/pdf-import-marker-normalization.js',
